refactor(studio/scrapbook): split pages/scrapbook.ts (809 → 17-line barrel + 6 modules) (#207)
Pure mechanical refactor — no behavior change, no public-API change.
Mirrors the established `pages/entry-review/` pattern: slim entry
point at `pages/scrapbook.ts` re-exports from `pages/scrapbook/index.ts`;
implementation lives under the sibling directory.

| Module | Lines | Role |
|---|---|---|
| `pages/scrapbook.ts` | 17 | Barrel — re-exports `renderScrapbookPage` + `ScrapbookPageError` |
| `pages/scrapbook/index.ts` | 121 | Orchestrator — `renderScrapbookPage`, dispatch + render composition, layout call |
| `pages/scrapbook/dispatch.ts` | 128 | `resolveListing`, `lookupEntryId`, `ScrapbookPageError` |
| `pages/scrapbook/render.ts` | 378 | All RawHtml emitters + `computeKindMeta`, `buildFileFetchUrl`, `KindCounts` |
| `pages/scrapbook/image-readers.ts` | 137 | Image-format helpers — `readImageDimensions`, `readJpegDimensions`, `readWebpDimensions` |
| `pages/scrapbook/text-helpers.ts` | 86 | Pure Buffer/string helpers — `escapeHtml`, `stripFrontmatter`, `previewExcerpt`, etc. |
| `pages/scrapbook/types.ts` | 27 | Shared `RenderCtx` interface |

All 7 files under the 300-line line of preference; largest is
render.ts at 378 (well under the 500-line cap).

**Public API unchanged** — `renderScrapbookPage` and
`ScrapbookPageError` keep the same export shape from
`pages/scrapbook.ts`. The single external import site
(`packages/studio/src/server.ts:47`) was not edited.

**Tests:** identical counts before and after (`@deskwork/studio` 412
passed / 11 skipped). No new tests; no test file modifications.

**Live verification:** restarted studio against the post-refactor dist;
all three key routes return 200:

- `/dev/scrapbook/<site>/<path>?entryId=<uuid>` — entry-aware viewer
- `/dev/editorial-review/entry/<uuid>` — review surface
- `/dev/editorial-studio` — dashboard

**Decisions not pre-specified:**

- Chose the `pages/entry-review/` pattern (slim re-export at
  `pages/scrapbook.ts` + implementation under `pages/scrapbook/`)
  over the post-#191 envelope/dispatch sibling-files-at-same-level
  pattern. The `pages/` subdirectory pattern is the established
  precedent for this part of the tree and co-locates the 6 new
  modules.
- Added a 6th module (`text-helpers.ts`) beyond the 3 the issue
  proposed. The pure Buffer/string helpers (`escapeHtml`,
  `stripFrontmatter`, `previewExcerpt`, `countLines`, `countJsonKeys`)
  are used only by `render.ts` but are independent of the render
  context — keeping them out of `render.ts` keeps that module focused
  on RawHtml emission.

Closes (pending v0.16.0 marketplace-walk verification): #207.
