+++ docs/0.16.0/001-IN-PROGRESS/open-issue-tranche-cleanup/README.md
@@ -21,7 +21,7 @@ Triage the open-issue tracker for the `deskwork`, `deskwork-studio`, and support
 | 5 | Sweep moot/superseded/stale (Tranche 4 + 5); reframe #40, #53; triage #92 | Done |
 | 6 | Tracker audit — verify the open list reflects only currently-actionable work | Not started |
 | 7 | Fix marginalia edit + delete UX (#199) | Text-edit + delete fix-landed; category-edit fix-landed (#204); range-edit wontfix (#203 closed) |
-| 8 | Extend entry-aware addressing to scrapbook viewer + link emitters (#205) | Fix-landed (pending v0.16.0 verification); #207 split filed |
+| 8 | Extend entry-aware addressing to scrapbook viewer + link emitters (#205) | Fix-landed (pending v0.16.0 verification); #207 split fix-landed |
 | 9 | Ingest defaults to Drafting per add/ingest semantic distinction (#206) | Not started |
 
 ## Key Links
