+++ packages/studio/src/pages/scrapbook.ts
@@ -1,809 +1,17 @@
 /**
  * Scrapbook viewer — `/dev/scrapbook/:site/<path>`.
  *
- * Issue #161 redesign: aside-left folder card with numbered item list,
- * vertical card grid with per-kind colored ribbons + always-visible foot
- * toolbar + per-kind preview rendering, drop zone, secret section,
- * single-expanded card invariant, aside cross-linking.
+ * The implementation lives in the `scrapbook/` directory (one module
+ * per concern, each under the 500-line cap). This file re-exports the
+ * public render function + error type so the existing `server.ts`
+ * import keeps working.
  *
- * Mockup: docs/superpowers/frontend-design/2026-05-02-review-redesign/scrapbook-redesign.html
- * Spec:   docs/superpowers/specs/2026-05-02-scrapbook-redesign-impl-spec.md
+ * Mirrors the post-#191 envelope/dispatch split for mutations
+ * (`routes/scrapbook-mutation-{envelope,dispatch}.ts`): the slim entry
+ * is the public surface; the directory holds the concerns.
  */
 
-import { existsSync, readFileSync } from 'node:fs';
-import { readCalendar } from '@deskwork/core/calendar';
-import { findEntry } from '@deskwork/core/calendar-mutations';
-import { resolveCalendarPath } from '@deskwork/core/paths';
-import {
-  formatRelativeTime,
-  formatSize,
-  listScrapbook,
-  listScrapbookForEntry,
-  scrapbookDirAtPath,
-  scrapbookDirForEntry,
-  scrapbookFilePathAtDir,
-  type ScrapbookItem,
-  type ScrapbookItemKind,
-  type ScrapbookSummary,
-} from '@deskwork/core/scrapbook';
-import { readSidecar } from '@deskwork/core/sidecar';
-import { UUID_RE } from '../routes/scrapbook-mutation-envelope.ts';
-import type { StudioContext } from '../routes/api.ts';
-import { html, unsafe, type RawHtml } from './html.ts';
-import { layout } from './layout.ts';
-import { renderEditorialFolio } from './chrome.ts';
-
-const KIND_LABEL: Record<ScrapbookItemKind, string> = {
-  md: 'MD',
-  img: 'IMG',
-  json: 'JSON',
-  js: 'JS',
-  txt: 'TXT',
-  other: '·',
-};
-
-function escapeHtml(s: string): string {
-  return s
-    .replace(/&/g, '&amp;')
-    .replace(/</g, '&lt;')
-    .replace(/>/g, '&gt;')
-    .replace(/"/g, '&quot;')
-    .replace(/'/g, '&#39;');
-}
-
-/**
- * Strip a YAML frontmatter block from the top of an md file. Only strips
- * the leading `---\n...\n---\n` block; body-level `---` separators (Setext
- * H2 underline, thematic break) are preserved because the function only
- * looks at the first 4 chars for the opener.
- */
-function stripFrontmatter(text: string): string {
-  if (!text.startsWith('---\n')) return text;
-  const closeIdx = text.indexOf('\n---\n', 4);
-  if (closeIdx < 0) return text;
-  return text.slice(closeIdx + 5).replace(/^\n+/, '');
-}
-
-/**
- * Build the closed-state preview excerpt for md/json/txt. Returns null
- * when there's nothing useful to render — empty file, frontmatter-only
- * file, or binary masquerading as text — so the caller can omit the
- * preview block entirely (matches "other" kind treatment, avoids the
- * 6rem min-height void).
- *
- * For json: pretty-print via JSON.parse + JSON.stringify(_, null, 2) so
- * minified single-line files still render multi-line. Falls back to raw
- * content on parse error (bad JSON is still readable as text).
- *
- * Binary detection: NUL byte presence after UTF-8 decode. Real text
- * almost never has NUL; binary files have it within the first KB.
- */
-function previewExcerpt(buf: Buffer, kind: 'md' | 'json' | 'txt'): string | null {
-  let text = buf.subarray(0, Math.min(buf.byteLength, 2400)).toString('utf-8');
-  if (text.indexOf('\0') >= 0) return null;
-  if (kind === 'md') text = stripFrontmatter(text);
-  if (kind === 'json') {
-    try {
-      const fullText = buf.toString('utf-8');
-      text = JSON.stringify(JSON.parse(fullText), null, 2);
-    } catch {
-      // Invalid JSON — fall through to the raw-text excerpt below.
-    }
-  }
-  const excerpt = text.split('\n').slice(0, 8).join('\n').slice(0, 600);
-  if (excerpt.trim() === '') return null;
-  return excerpt;
-}
-
-/**
- * Count lines in a text file: number of `\n` bytes plus 1 if the last
- * byte isn't `\n` (so a 3-line file whether or not it has a trailing
- * newline reports 3).
- */
-function countLines(buf: Buffer): number {
-  let count = 0;
-  for (const b of buf) if (b === 0x0a) count++;
-  if (buf.length > 0 && buf[buf.length - 1] !== 0x0a) count++;
-  return count;
-}
-
-/**
- * Count top-level keys in a JSON object. Returns null if the file is not
- * valid JSON or its root is not a plain object (arrays, primitives →
- * null; caller renders no extra meta).
- */
-function countJsonKeys(buf: Buffer): number | null {
-  try {
-    const obj: unknown = JSON.parse(buf.toString('utf-8'));
-    if (obj !== null && typeof obj === 'object' && !Array.isArray(obj)) {
-      return Object.keys(obj).length;
-    }
-    return null;
-  } catch {
-    return null;
-  }
-}
-
-interface ImageDimensions { readonly width: number; readonly height: number; }
-
-/**
- * Read width × height from a buffer. Recognizes PNG, JPEG, WebP, and
- * GIF; returns null for unrecognized signatures or truncated/malformed
- * buffers. Used by the scrapbook card meta to render `{W} × {H}` next
- * to the image kind chip.
- *
- * Each format is parsed from its file-header structure (no external
- * dependency); each branch returns null on any unexpected byte rather
- * than throwing, so a corrupt image still renders with empty meta.
- */
-function readImageDimensions(buf: Buffer): ImageDimensions | null {
-  if (buf.length < 12) return null;
-  // PNG: 89 50 4E 47 0D 0A 1A 0A then IHDR (chunk-length 4, "IHDR" 4,
-  // width 4, height 4 — width@16, height@20).
-  if (
-    buf[0] === 0x89 && buf[1] === 0x50 && buf[2] === 0x4e && buf[3] === 0x47
-  ) {
-    if (buf.length < 24) return null;
-    return { width: buf.readUInt32BE(16), height: buf.readUInt32BE(20) };
-  }
-  // GIF: "GIF87a" or "GIF89a" + logical screen descriptor (width LE @6,
-  // height LE @8).
-  if (
-    buf[0] === 0x47 && buf[1] === 0x49 && buf[2] === 0x46 &&
-    buf[3] === 0x38 && (buf[4] === 0x37 || buf[4] === 0x39) && buf[5] === 0x61
-  ) {
-    if (buf.length < 10) return null;
-    return { width: buf.readUInt16LE(6), height: buf.readUInt16LE(8) };
-  }
-  // JPEG: starts FF D8. Width/height live in a Start-Of-Frame marker
-  // (FF C0–CF, excluding DHT C4 / JPG C8 / DAC CC). Walk markers
-  // skipping their payloads until the SOF is found.
-  if (buf[0] === 0xff && buf[1] === 0xd8) {
-    return readJpegDimensions(buf);
-  }
-  // WebP: "RIFF" {size} "WEBP" then a chunk (VP8 / VP8L / VP8X).
-  if (
-    buf[0] === 0x52 && buf[1] === 0x49 && buf[2] === 0x46 && buf[3] === 0x46 &&
-    buf[8] === 0x57 && buf[9] === 0x45 && buf[10] === 0x42 && buf[11] === 0x50
-  ) {
-    return readWebpDimensions(buf);
-  }
-  return null;
-}
-
-function readJpegDimensions(buf: Buffer): ImageDimensions | null {
-  // Marker walk: skip FF D8 (SOI), then each marker is FF Xn followed
-  // by a 2-byte big-endian segment length (which includes its own 2
-  // bytes). The SOF segment's payload is: 1 byte precision, 2 bytes
-  // height, 2 bytes width (the rest is component info).
-  let i = 2;
-  while (i + 4 <= buf.length) {
-    if (buf[i] !== 0xff) return null;
-    let marker = buf[i + 1] ?? 0;
-    // Skip fill bytes (0xff padding before the actual marker byte).
-    while (marker === 0xff && i + 2 < buf.length) {
-      i++;
-      marker = buf[i + 1] ?? 0;
-    }
-    i += 2;
-    // Standalone markers (no length): RST0–7 (D0–D7) and SOI/EOI/TEM.
-    if (marker === 0xd9 || marker === 0xd8 || marker === 0x01) return null;
-    if (marker >= 0xd0 && marker <= 0xd7) continue;
-    if (i + 2 > buf.length) return null;
-    const segLen = buf.readUInt16BE(i);
-    // SOF markers carry the dimensions. Exclusions per JPEG spec:
-    // C4 (DHT — Huffman tables), C8 (JPG reserved), CC (DAC — arithmetic
-    // coding conditioning).
-    const isSof =
-      marker >= 0xc0 && marker <= 0xcf &&
-      marker !== 0xc4 && marker !== 0xc8 && marker !== 0xcc;
-    if (isSof) {
-      if (i + 7 > buf.length) return null;
-      return {
-        width: buf.readUInt16BE(i + 5),
-        height: buf.readUInt16BE(i + 3),
-      };
-    }
-    i += segLen;
-  }
-  return null;
-}
-
-function readWebpDimensions(buf: Buffer): ImageDimensions | null {
-  // Three sub-formats (per RFC 6386 / VP8L spec / WebP container spec):
-  //   - "VP8 " (lossy):     header @20-22: signature 9D 01 2A, then
-  //                          14-bit width LE @23, 14-bit height LE @25
-  //                          (each masked to 14 bits — top 2 bits are
-  //                          horizontal/vertical scale).
-  //   - "VP8L" (lossless):  header @20: signature 0x2F, then 32 bits
-  //                          packed: width-1 (14 bits) + height-1 (14
-  //                          bits) + alpha-flag + version (4 bits).
-  //   - "VP8X" (extended):  flags @20, reserved @21-23, width-1 (24
-  //                          bits LE) @24, height-1 (24 bits LE) @27.
-  // Need at least the RIFF header + chunk fourcc; per-variant length
-  // checks below cover the variant-specific payload sizes.
-  if (buf.length < 16) return null;
-  const fourcc = buf.subarray(12, 16).toString('ascii');
-  if (fourcc === 'VP8 ') {
-    if (buf.length < 30) return null;
-    if (buf[23] !== 0x9d || buf[24] !== 0x01 || buf[25] !== 0x2a) return null;
-    const w = buf.readUInt16LE(26) & 0x3fff;
-    const h = buf.readUInt16LE(28) & 0x3fff;
-    return { width: w, height: h };
-  }
-  if (fourcc === 'VP8L') {
-    if (buf.length < 25) return null;
-    if (buf[20] !== 0x2f) return null;
-    // Read 4 bytes LE at offset 21, then unpack.
-    const packed = buf.readUInt32LE(21);
-    const widthMinus1 = packed & 0x3fff;
-    const heightMinus1 = (packed >>> 14) & 0x3fff;
-    return { width: widthMinus1 + 1, height: heightMinus1 + 1 };
-  }
-  if (fourcc === 'VP8X') {
-    if (buf.length < 30) return null;
-    // 24-bit little-endian: low byte + (mid << 8) + (high << 16).
-    const w =
-      ((buf[24] ?? 0) | ((buf[25] ?? 0) << 8) | ((buf[26] ?? 0) << 16)) + 1;
-    const h =
-      ((buf[27] ?? 0) | ((buf[28] ?? 0) << 8) | ((buf[29] ?? 0) << 16)) + 1;
-    return { width: w, height: h };
-  }
-  return null;
-}
-
-/**
- * Per-render context for the scrapbook page. Carries a pre-resolved
- * absolute scrapbook directory (used for filesystem reads in
- * `computeKindMeta` / `renderPreview`) plus the addressing identifiers
- * used in URL emission (`site`, `path`, optional `entryId`).
- *
- * #205 — the entryId is the load-bearing addition. When present, it
- * threads into the file-fetch URL via `entryId=` so the read-only
- * binary endpoint resolves through `scrapbookDirForEntry` (matching
- * the entry-aware mutation API). When absent, slug-template addressing
- * (`path=`) is the back-compat fallback.
- */
-interface RenderCtx {
-  studio: StudioContext;
-  site: string;
-  path: string;
-  entryId?: string;
-  /** Absolute path of the scrapbook directory on disk. */
-  scrapbookDir: string;
-}
-
-/**
- * Compute the per-kind extra meta string shown after the kind chip + size:
- *   md / txt → "{N} lines"
- *   json     → "{N} keys" (root must be a plain object; otherwise empty)
- *   img      → "{W} × {H}" (PNG only; other formats → empty)
- *   other    → empty
- *
- * ENOENT (race-window with delete) returns empty so the card still
- * renders; other errors propagate to the page renderer.
- */
-function computeKindMeta(
-  rctx: RenderCtx,
-  item: ScrapbookItem,
-  opts: { secret?: boolean } = {},
-): string {
-  if (item.kind !== 'md' && item.kind !== 'txt' && item.kind !== 'json' && item.kind !== 'img') {
-    return '';
-  }
-  let buf: Buffer;
-  try {
-    const fullPath = scrapbookFilePathAtDir(
-      rctx.scrapbookDir,
-      item.name,
-      opts.secret ? { secret: true } : {},
-    );
-    buf = readFileSync(fullPath);
-  } catch (e) {
-    if (e instanceof Error && 'code' in e && e.code === 'ENOENT') return '';
-    throw e;
-  }
-  if (item.kind === 'md' || item.kind === 'txt') return `${countLines(buf)} lines`;
-  if (item.kind === 'json') {
-    const keys = countJsonKeys(buf);
-    return keys !== null ? `${keys} keys` : '';
-  }
-  // img
-  const dims = readImageDimensions(buf);
-  return dims ? `${dims.width} × ${dims.height}` : '';
-}
-
-/**
- * Build the file-fetch URL for the read-only binary endpoint. Prefers
- * entry-aware addressing when `rctx.entryId` is set (#205); falls back
- * to slug-template addressing (`path=`) otherwise.
- */
-function buildFileFetchUrl(
-  rctx: RenderCtx,
-  filename: string,
-  secret: boolean,
-): string {
-  const params = new URLSearchParams({ site: rctx.site, name: filename });
-  if (rctx.entryId !== undefined && rctx.entryId.length > 0) {
-    params.set('entryId', rctx.entryId);
-  } else {
-    params.set('path', rctx.path);
-  }
-  if (secret) params.set('secret', '1');
-  return `/api/dev/scrapbook-file?${params.toString()}`;
-}
-
-/**
- * Server-side preview for the closed-state card. Img → bg-frame URL;
- * md → italic Newsreader excerpt with frontmatter stripped; json → mono
- * pre with parse-then-stringify pretty-print; txt → mono pre raw excerpt.
- * Other / empty / binary-as-text → no preview block.
- */
-function renderPreview(
-  rctx: RenderCtx,
-  item: ScrapbookItem,
-  opts: { secret?: boolean } = {},
-): RawHtml {
-  const { secret = false } = opts;
-  if (item.kind === 'img') {
-    const url = buildFileFetchUrl(rctx, item.name, secret);
-    return unsafe(html`
-      <div class="scrap-preview scrap-preview--img" aria-hidden="true">
-        <div class="scrap-preview--img-frame" style="background-image: url(&quot;${url}&quot;);"></div>
-      </div>`);
-  }
-  if (item.kind !== 'md' && item.kind !== 'txt' && item.kind !== 'json') {
-    return unsafe('');
-  }
-  try {
-    const fullPath = scrapbookFilePathAtDir(
-      rctx.scrapbookDir,
-      item.name,
-      secret ? { secret: true } : {},
-    );
-    const buf = readFileSync(fullPath);
-    const excerpt = previewExcerpt(buf, item.kind);
-    if (excerpt === null) return unsafe('');
-    const safe = escapeHtml(excerpt);
-    if (item.kind === 'json' || item.kind === 'txt') {
-      return unsafe(html`
-        <pre class="scrap-preview scrap-preview--mono" aria-hidden="true">${unsafe(safe)}</pre>`);
-    }
-    return unsafe(html`
-      <div class="scrap-preview scrap-preview-md" aria-hidden="true"><p>${unsafe(safe)}</p></div>`);
-  } catch (e) {
-    // ENOENT = file disappeared between listScrapbook and this read (race
-    // window with delete); rendering an empty preview is the right call.
-    // Anything else (EACCES, EISDIR, encoding bugs) propagates so the
-    // operator sees a real error instead of a silently-broken page.
-    if (e instanceof Error && 'code' in e && e.code === 'ENOENT') {
-      return unsafe('');
-    }
-    throw e;
-  }
-}
-
-interface KindCounts {
-  all: number;
-  md: number;
-  img: number;
-  json: number;
-  js: number;
-  txt: number;
-  other: number;
-}
-
-function countByKind(items: readonly ScrapbookItem[]): KindCounts {
-  const counts: KindCounts = {
-    all: items.length,
-    md: 0,
-    img: 0,
-    json: 0,
-    js: 0,
-    txt: 0,
-    other: 0,
-  };
-  for (const i of items) counts[i.kind]++;
-  return counts;
-}
-
-function renderFilterChips(counts: KindCounts): RawHtml {
-  const chip = (kind: keyof KindCounts, label: string, isAll = false): RawHtml =>
-    unsafe(html`
-    <button class="scrap-filter" type="button" data-filter="${kind}"
-      aria-pressed="${isAll ? 'true' : 'false'}">${label} · ${counts[kind]}</button>`);
-  return unsafe(html`
-    <div class="scrap-filters" role="toolbar" aria-label="filter by kind">
-      ${chip('all', 'all', true)}
-      ${chip('md', 'md')}
-      ${chip('img', 'img')}
-      ${chip('json', 'json')}
-      ${chip('txt', 'txt')}
-      ${chip('other', 'other')}
-    </div>`);
-}
-
-function renderSearch(): RawHtml {
-  return unsafe(html`
-    <div class="scrap-search">
-      <input type="search" placeholder="filter by name or content" aria-label="filter scrapbook" data-scrap-search />
-      <span class="scrap-search-kbd">/</span>
-    </div>`);
-}
-
-function renderBreadcrumb(site: string, path: string): RawHtml {
-  const segments = path.split('/').filter(Boolean);
-  const last = segments[segments.length - 1] ?? path;
-  return unsafe(html`
-    <nav class="scrap-breadcrumb" aria-label="hierarchy">
-      <a href="/dev/content/${site}">${site}</a><span class="sep">›</span>
-      <b>${last}</b>
-    </nav>`);
-}
-
-function renderAside(
-  site: string,
-  path: string,
-  items: readonly ScrapbookItem[],
-  totalSize: number,
-  lastModified: string | null,
-  secretCount: number,
-  reviewLink: string | null,
-): RawHtml {
-  const lastModifiedLabel = lastModified ? formatRelativeTime(lastModified) : '—';
-  const publicCount = items.length;
-  const sizeLabel = formatSize(totalSize);
-  const folderLabel = path.split('/').filter(Boolean).pop() ?? path;
-  const fullPath = `${site}/${path}/scrapbook/`;
-  // #168 Phase 34 ship-pass — when this scrapbook belongs to a tracked
-  // calendar entry, expose a "← back to review" link so the operator
-  // who arrived from the entry-review surface (or via the dashboard's
-  // scrapbook chip) has an obvious path back. Pre-fix the only nav
-  // affordance was the breadcrumb's site link, which lands on the
-  // content tree, not the entry-review.
-  const backLink: RawHtml = reviewLink !== null
-    ? unsafe(html`<p class="scrap-aside-back"><a href="${reviewLink}">← back to review</a></p><hr />`)
-    : unsafe('');
-  return unsafe(html`
-    <aside class="scrap-aside">
-      ${backLink}
-      <p class="scrap-aside-kicker"><em>§</em> The folder</p>
-      <h1 class="scrap-aside-title">${folderLabel}</h1>
-      <p class="scrap-aside-meta">${site}</p>
-      <hr />
-      <p class="scrap-aside-totals">
-        <strong>${publicCount}</strong> public ·
-        <strong>${secretCount}</strong> secret ·
-        <em>${sizeLabel}</em>
-      </p>
-      <p class="scrap-aside-meta">last modified ${lastModifiedLabel}</p>
-      <hr />
-      <ol class="scrap-aside-list" data-scrap-aside-list>
-        ${items.map((item, i) => {
-          const seq = String(i + 1).padStart(2, '0');
-          return unsafe(html`<li><span class="num">${seq}</span><a href="#item-${i + 1}" data-scrap-aside-link>${item.name}</a></li>`);
-        })}
-      </ol>
-      <hr />
-      <div class="scrap-aside-actions">
-        <button class="scrap-aside-btn scrap-aside-btn--primary" type="button" data-action="new-note">+ new note</button>
-        <button class="scrap-aside-btn" type="button" data-action="upload">+ upload file</button>
-      </div>
-      <hr />
-      <p class="scrap-aside-path">${fullPath}</p>
-    </aside>`);
-}
-
-function renderCard(
-  rctx: RenderCtx,
-  item: ScrapbookItem,
-  index: number,
-  opts: { secret?: boolean } = {},
-): RawHtml {
-  const { secret = false } = opts;
-  const seq = String(index + 1).padStart(2, '0');
-  const kindLabel = KIND_LABEL[item.kind];
-  const kindClass = item.kind === 'other' ? '' : `scrap-kind--${item.kind}`;
-  const time = item.mtime
-    ? html`<time class="scrap-time" datetime="${item.mtime}">${formatRelativeTime(item.mtime)}</time>`
-    : '';
-  const preview = renderPreview(rctx, item, { secret });
-  const kindMeta = computeKindMeta(rctx, item, { secret });
-  const kindMetaHtml: RawHtml = kindMeta
-    ? unsafe(html`<span>·</span><span>${kindMeta}</span>`)
-    : unsafe('');
-  const editBtn = item.kind === 'img'
-    ? unsafe('')
-    : unsafe(html`<button class="scrap-tool" type="button" data-action="edit">edit</button>`);
-  // Secret cards get id="secret-item-N" to disambiguate from public ids in
-  // restoreFromHash + aside cross-link lookups (F4 contract); the
-  // mark-secret action toggle reads "mark public" since clicking it moves
-  // the card OUT of the secret section.
-  const id = secret ? `secret-item-${index + 1}` : `item-${index + 1}`;
-  const markSecretLabel = secret ? 'mark public' : 'mark secret';
-  const dataSecretAttr = secret ? ' data-secret="true"' : '';
-  // #164 Phase 34b — small ⚿ glyph next to .scrap-name on secret
-  // cards. Provides visual continuity for the secret marker once a
-  // card is expanded (where it grows outside the grouped section's
-  // visual scope).
-  const secretGlyph: RawHtml = secret
-    ? unsafe(html`<span class="scrap-name-secret-mark" aria-label="secret" title="secret — never published">⚿</span>`)
-    : unsafe('');
-  return unsafe(html`
-    <li class="scrap-card" data-kind="${item.kind}" data-state="closed" id="${id}"${unsafe(dataSecretAttr)}>
-      <div class="scrap-card-head">
-        <span class="scrap-seq">N° ${seq}</span>
-        ${secretGlyph}
-        <span class="scrap-name" data-action="open">${item.name}</span>
-        ${unsafe(time)}
-      </div>
-      <div class="scrap-card-meta">
-        <span class="scrap-kind ${kindClass}">${kindLabel}</span>
-        <span class="scrap-size">${formatSize(item.size)}</span>
-        ${kindMetaHtml}
-      </div>
-      ${preview}
-      <div class="scrap-card-foot">
-        <button class="scrap-tool scrap-tool--primary" type="button" data-action="open">open</button>
-        ${editBtn}
-        <button class="scrap-tool" type="button" data-action="rename">rename</button>
-        <button class="scrap-tool" type="button" data-action="mark-secret">${markSecretLabel}</button>
-        <span class="spacer"></span>
-        <button class="scrap-tool scrap-tool--delete" type="button" data-action="delete">delete</button>
-      </div>
-    </li>`);
-}
-
-/**
- * Inline new-note composer (Phase 34b — #166).
- *
- * Mirrors the pre-F1 inline composer (`44094ee^:scrapbook.ts:274-294`),
- * adapted to the F1 `.scrap-*` design vocabulary. Hidden by default;
- * the aside's `+ new note` button reveals it via the client wire-up.
- *
- * Per `.claude/rules/affordance-placement.md`: component-attached to
- * the page (not a generic toolbar), placed where the resulting note
- * will appear in sorted position. Direct manipulation: in-page form,
- * filename + body + secret toggle visible inline, Cmd/Ctrl+S saves,
- * Esc cancels. Replaces the F1 `window.prompt()` regression (#166).
- */
-function renderComposer(): RawHtml {
-  return unsafe(html`
-    <form class="scrap-composer" data-scrap-composer hidden>
-      <header class="scrap-composer-head">
-        <span class="scrap-composer-glyph" aria-hidden="true">✎</span>
-        <span class="scrap-composer-kicker">NEW NOTE</span>
-        <input type="text" class="scrap-composer-filename" data-composer-filename
-          placeholder="note-name.md" aria-label="new note filename" />
-        <label class="scrap-composer-secret" title="save under scrapbook/secret/ — never published">
-          <input type="checkbox" data-composer-secret />
-          <span>secret</span>
-        </label>
-        <button class="scrap-tool" type="button" data-action="composer-cancel">cancel</button>
-        <button class="scrap-tool scrap-tool--primary" type="submit" data-action="composer-save">save →</button>
-      </header>
-      <textarea class="scrap-composer-body" data-composer-body
-        placeholder="Write the note in markdown. Cmd/Ctrl+S saves, Esc cancels."
-        aria-label="new note body" rows="8"></textarea>
-    </form>`);
-}
-
-function renderDropZone(): RawHtml {
-  return unsafe(html`
-    <div class="scrap-drop" role="button" tabindex="0" data-action="upload"
-         aria-label="Drop a file here, or press Enter to pick one">
-      ── drop a file here, or pick one ──
-    </div>`);
-}
-
-function renderSecretSection(
-  rctx: RenderCtx,
-  secretItems: readonly ScrapbookItem[],
-): RawHtml {
-  if (secretItems.length === 0) return unsafe('');
-  const cards = secretItems.map((item, i) => renderCard(rctx, item, i, { secret: true }));
-  return unsafe(html`
-    <section class="scrap-secret" aria-label="secret items">
-      <header class="scrap-secret-head">
-        <span class="scrap-secret-mark" aria-hidden="true">⚿</span>
-        <h2 class="scrap-secret-title">Secret</h2>
-        <span class="scrap-secret-badge">private — never published</span>
-      </header>
-      <ol class="scrap-cards">
-        ${cards}
-      </ol>
-    </section>`);
-}
-
-/**
- * #168 Phase 34 ship-pass — when the scrapbook path matches a tracked
- * calendar entry with a stamped UUID, return the entry's id so the aside
- * can render a "← back to review" link AND the client can address
- * mutations via `entryId` (#191 fix). Returns null when no entry matches
- * (organizational subdirs, ad-hoc paths, or pre-doctor entries lacking
- * an id) — the link is then omitted and mutations fall back to slug-
- * template addressing.
- *
- * Failures (calendar absent, parse error) fall through to null so a
- * transient calendar issue never blocks the scrapbook render.
- */
-function lookupEntryId(
-  ctx: StudioContext,
-  site: string,
-  path: string,
-): string | null {
-  if (!(site in ctx.config.sites)) return null;
-  try {
-    const calendarPath = resolveCalendarPath(ctx.projectRoot, ctx.config, site);
-    if (!existsSync(calendarPath)) return null;
-    const cal = readCalendar(calendarPath);
-    const entry = findEntry(cal, path);
-    if (!entry || !entry.id) return null;
-    return entry.id;
-  } catch {
-    return null;
-  }
-}
-
-/**
- * Error class surfaced by the page route to translate into HTTP status
- * codes (`server.ts` maps these onto 400/404 responses).
- */
-export class ScrapbookPageError extends Error {
-  readonly status: 400 | 404;
-  constructor(message: string, status: 400 | 404) {
-    super(message);
-    this.status = status;
-    this.name = 'ScrapbookPageError';
-  }
-}
-
-/**
- * #205 — listing + dir-resolution dispatch. Two addressing modes:
- *
- *   - Entry-id mode: `entryId` is validated as a UUID, the sidecar is
- *     read, and the listing resolves via `listScrapbookForEntry` so
- *     non-kebab-case entries (feature-doc layouts, dotted version
- *     segments, etc.) read items from the entry's artifact-parent
- *     `scrapbook/` directory. Symmetric to the entry-aware mutation API.
- *   - Slug mode: legacy / organizational paths under the site's content
- *     dir. Resolves via `scrapbookDirAtPath` + `listScrapbook` (the
- *     pre-#205 behavior).
- *
- * Returns the resolved `scrapbookDir` (absolute) plus the listing
- * summary so the caller threads both through to render helpers.
- */
-async function resolveListing(
-  ctx: StudioContext,
-  site: string,
-  path: string,
-  entryId: string | null,
-): Promise<{ scrapbookDir: string; result: ScrapbookSummary; resolvedEntryId: string | null }> {
-  if (entryId !== null) {
-    if (!UUID_RE.test(entryId)) {
-      throw new ScrapbookPageError(`invalid entryId`, 400);
-    }
-    let entry;
-    try {
-      entry = await readSidecar(ctx.projectRoot, entryId);
-    } catch (e) {
-      // readSidecar throws "sidecar not found" with ENOENT; map to 404.
-      const reason = e instanceof Error ? e.message : String(e);
-      if (/sidecar not found/i.test(reason)) {
-        throw new ScrapbookPageError(reason, 404);
-      }
-      throw e;
-    }
-    const scrapbookDir = scrapbookDirForEntry(
-      ctx.projectRoot,
-      ctx.config,
-      site,
-      { id: entry.uuid, slug: entry.slug },
-    );
-    const result = listScrapbookForEntry(
-      ctx.projectRoot,
-      ctx.config,
-      site,
-      { id: entry.uuid, slug: entry.slug },
-    );
-    return { scrapbookDir, result, resolvedEntryId: entry.uuid };
-  }
-  // Slug-mode: legacy back-compat. listScrapbook returns
-  // { exists: false, items: [] } for missing dirs, so an empty
-  // scrapbook is not an error path. Real errors (slug validation,
-  // scrapbookDir resolution failures, FS permission issues)
-  // propagate to the studio's error handler.
-  const scrapbookDir = scrapbookDirAtPath(ctx.projectRoot, ctx.config, site, path);
-  const result = listScrapbook(ctx.projectRoot, ctx.config, site, path);
-  // Best-effort lookup so the back-link to the review surface still
-  // renders for slug-template entries that happen to match a calendar
-  // entry by slug.
-  const resolvedEntryId = lookupEntryId(ctx, site, path);
-  return { scrapbookDir, result, resolvedEntryId };
-}
-
-export async function renderScrapbookPage(
-  ctx: StudioContext,
-  site: string,
-  path: string,
-  opts: { entryId?: string } = {},
-): Promise<string> {
-  // Validate site against the project's configured site list. Without
-  // this check, an unknown site key reaches the path resolver and
-  // produces either an opaque error or a path traversal vector.
-  if (!(site in ctx.config.sites)) {
-    throw new ScrapbookPageError(`unknown site: ${site}`, 404);
-  }
-  const requestedEntryId =
-    opts.entryId !== undefined && opts.entryId.length > 0 ? opts.entryId : null;
-  const { scrapbookDir, result, resolvedEntryId } = await resolveListing(
-    ctx,
-    site,
-    path,
-    requestedEntryId,
-  );
-  const items = result.items;
-  const secretItems = result.secretItems;
-  const totalSize = items.reduce((s, i) => s + i.size, 0);
-  const lastModified = items.reduce<string | null>((acc, i) => {
-    if (!i.mtime) return acc;
-    if (!acc || i.mtime > acc) return i.mtime;
-    return acc;
-  }, null);
-  const counts = countByKind(items);
-  const folderLabel = path.split('/').filter(Boolean).pop() ?? path;
-  // Effective entryId for URL emission: prefer the request-supplied id
-  // (so the client's mutation requests round-trip through the same
-  // addressing mode) and fall back to the slug-mode lookup result.
-  const effectiveEntryId = requestedEntryId ?? resolvedEntryId;
-  const rctx: RenderCtx = effectiveEntryId !== null
-    ? { studio: ctx, site, path, entryId: effectiveEntryId, scrapbookDir }
-    : { studio: ctx, site, path, scrapbookDir };
-  const cards = items.map((item, i) => renderCard(rctx, item, i));
-  const cardsHtml = cards.map((c) => c.__raw).join('');
-  const reviewLink =
-    effectiveEntryId !== null
-      ? `/dev/editorial-review/entry/${effectiveEntryId}`
-      : null;
-  // The data-entry-id attribute is consumed by scrapbook-client.ts —
-  // when present, the client sends `entryId` on mutation requests so
-  // writes resolve via `scrapbookDirForEntry` (#191). Falls back to
-  // `data-path` slug-template addressing when absent.
-  // entryId comes from a UUID lookup — already validated against the
-  // calendar's CalendarEntry.id, so no escaping concern beyond the
-  // belt-and-braces unsafe wrapping for the conditional attribute.
-  const entryIdAttr =
-    effectiveEntryId !== null
-      ? unsafe(` data-entry-id="${effectiveEntryId}"`)
-      : unsafe('');
-  const body = html`
-    ${renderEditorialFolio('content', `scrapbook · ${site}/${path}`)}
-    <main class="scrap-page" data-site="${site}" data-path="${path}"${entryIdAttr}>
-      ${renderAside(site, path, items, totalSize, lastModified, secretItems.length, reviewLink)}
-      <section class="scrap-main">
-        <header class="scrap-main-header">
-          ${renderBreadcrumb(site, path)}
-          ${renderSearch()}
-        </header>
-        ${renderFilterChips(counts)}
-        ${renderComposer()}
-        <ol class="scrap-cards" id="cards" data-scrap-cards>
-          ${unsafe(cardsHtml)}
-        </ol>
-        ${renderDropZone()}
-        ${renderSecretSection(rctx, secretItems)}
-      </section>
-    </main>`;
-  return layout({
-    title: `scrapbook · ${folderLabel} — dev`,
-    cssHrefs: [
-      '/static/css/editorial-review.css',
-      '/static/css/editorial-nav.css',
-      '/static/css/scrapbook.css',
-      '/static/css/blog-figure.css',
-    ],
-    bodyAttrs: 'data-review-ui="scrapbook"',
-    bodyHtml: body,
-    scriptModules: ['scrapbook-client'],
-  });
-}
+export {
+  renderScrapbookPage,
+  ScrapbookPageError,
+} from './scrapbook/index.ts';
